run: Allow fuzzy matching for apps when possible
Currently, the user has to provide at minimum the exact full application
id in order to run it. If they don't remember it, they have to first
find it using other commands or in other places.

This makes it harder for users to use, as they often need to do extra
steps before they can run their apps.

To fix that, allow fuzzy matching for apps when the provided partial ref
does not contain any dots or slashes, similarly to when installing.
There's already much of the code implemented for runtimes, so reuse that
and make the appriopriate variable names generic.